--- engine/reflection/private/instance.hpp
@@ -1,2 +1,0 @@
-#pragma once
-#include <impls.hpp>
