+++ engine/reflection/private/member/field_impl.hpp
@@ -1,6 +1,10 @@
 #pragma once
 #include <member/field.hpp>
 
+#include <instance/instance.hpp>
+#include <type/store.hpp>
+
+
 template <typename Class, typename Member>
 Field::Field(TypeStore& type_store, Member Class::* mem_ptr)
 {
