+++ engine/reflection/private/type/type.hpp
@@ -1,9 +1,11 @@
 #pragma once
 
 #include <common.hpp>
+
 #include <memory>
 #include <optional>
 #include <string>
+#include <string_view>
 #include <typeindex>
 #include <unordered_map>
 #include <vector>
@@ -11,27 +13,12 @@
 #include <utility/argument_list.hpp>
 #include <utility/parameter_list.hpp>
 
+class Constructor;
+class TypeStore;
+class Instance;
+class Destructor;
 class Field;
 class Method;
-class Constructor;
-class Destructor;
-class Constant;
-class Instance;
-
-class TypeStore
-{
-public:
-    TypeStore() = default;
-    ~TypeStore() = default;
-
-    NON_COPYABLE(TypeStore);
-
-    template <typename T> NO_DISCARD Type* get();
-    template <typename T> NO_DISCARD Instance makeInstance(T&& val);
-
-private:
-    std::unordered_map<std::type_index, std::unique_ptr<Type>> type_map;
-};
 
 class Type
 {
@@ -40,6 +27,7 @@ public:
 
     NON_COPYABLE(Type);
     DEFAULT_MOVABLE(Type);
+    ~Type();
 
     NO_DISCARD Instance construct(const ArgumentList& args) const;
     NO_DISCARD std::optional<uint64_t> getConstant(std::string_view name) const;
@@ -47,7 +35,7 @@ public:
     NO_DISCARD std::type_index getIndex() const;
 
 private:
-    Type(TypeStore* owner_store);
+    Type();
 
     // Always initialized
     TypeStore* owner_store {};
@@ -69,32 +57,4 @@ private:
     template <typename T> friend class TypeBuilder;
 };
 
-class Instance
-{
-public:
-    Instance() = default;
-
-    /// Instance from raw parts (unsafe)
-    Instance(void* ptr, const Type* type)
-        : ptr(ptr)
-        , type(type) { };
-
-    template <typename T> Instance(const Type* type_store, T val);
-
-    NON_COPYABLE(Instance);
-    Instance(Instance&&) noexcept;
-    Instance& operator=(Instance&&) noexcept;
-    ~Instance();
-
-    template <typename T> const T* cast() const;
-
-    template <typename T> T* cast();
-
-    NO_DISCARD bool isValid() const;
-    NO_DISCARD const Type* getType() const;
-    NO_DISCARD const void* getAddress() const;
-
-private:
-    void* ptr {};
-    const Type* type {};
-};
+template <typename T> bool Type::is() const { return this->getIndex() == typeid(T); }
