+++ engine/reflection/private/utility/parameter_list.cpp
@@ -1,8 +1,6 @@
 #include <utility/argument_list.hpp>
 #include <utility/parameter_list.hpp>
 
-#include <instance.hpp>
-
 NO_DISCARD bool ParameterList::validateArgs(const ArgumentList& args) const
 {
     auto as_types = args.asTypes();
